Updated tables in Statistics

+++ src/main/java/com/example/cloudanalystt/utils/ServersResponseTime.java
@@ -9,7 +9,7 @@ public class ServersResponseTime {
     private StringProperty max;
     private StringProperty avg;
 
-    public ServersResponseTime(String server, float min, float max, float avg) {
+    public ServersResponseTime(String server, double min, double max, double avg) {
         this.server = new SimpleStringProperty(server);
         this.min = new SimpleStringProperty(String.valueOf(min));
         this.max = new SimpleStringProperty(String.valueOf(max));
@@ -28,39 +28,39 @@ public class ServersResponseTime {
         this.server.set(server);
     }
 
-    public float getMin() {
-        return Float.parseFloat(min.get());
+    public double getMin() {
+        return Double.parseDouble(min.get());
     }
 
     public StringProperty minProperty() {
         return min;
     }
 
-    public void setMin(float min) {
+    public void setMin(double min) {
         this.min.set(String.valueOf(min));
     }
 
-    public float getMax() {
-        return Float.parseFloat(max.get());
+    public double getMax() {
+        return Double.parseDouble(max.get());
     }
 
     public StringProperty maxProperty() {
         return max;
     }
 
-    public void setMax(float max) {
+    public void setMax(double max) {
         this.max.set(String.valueOf(max));
     }
 
-    public float getAvg() {
-        return Float.parseFloat(avg.get());
+    public double getAvg() {
+        return Double.parseDouble(avg.get());
     }
 
     public StringProperty avgProperty() {
         return avg;
     }
 
-    public void setAvg(float avg) {
+    public void setAvg(double avg) {
         this.avg.set(String.valueOf(avg));
     }
 }
